Fix inefficient and insufficient regex for WWW-Authenticate
The AbstractBasicAuthHandler class of the urllib.request module uses
an inefficient regular expression which can be exploited by an
attacker to cause a denial of service. Fix the regex to prevent the
catastrophic backtracking.

Note that the original regex was roughly O(2**n)
The search for commas and spaces is unnecessary
(and insufficient to ensure that this starts a new scheme).
Replace with a simpler search for a word boundary, but also
only match where the scheme is basic, since the way this is used
only reads the realm when the scheme == 'basic'

Vulnerability reported by Matt Schwager.